Project/cloud_run/execute_script_service/requirements.txt
Flask==1.1.1
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.0.4
google-cloud-storage
marshmallow-dataclass == 7.2.0
marshmallow